authorization: gate and policies

+++ app/Http/Controllers/PostsController.php
@@ -4,10 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
+use Illuminate\Support\Facades\Gate;
 
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')
+            ->only(['create', 'store', 'edit', 'update', 'destroy']);
+    }
      /**
      * Display a listing of the resource.
      *
@@ -27,6 +34,7 @@ class PostsController extends Controller
      */
     public function create()
     {
+       // $this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -54,7 +62,9 @@ class PostsController extends Controller
     public function show($id)
     {
         //abort_if(!isset($this->posts[$id]), 404);
-        return view('posts.show', ['post' => BlogPost::findOrFail($id)]);
+        return view('posts.show', [
+            'post' => BlogPost::with('comments')->findOrFail($id)
+        ]);
     }
 
     /**
@@ -65,7 +75,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+        $this->authorize( $post);
+
+       /* if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }*/
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -78,6 +95,13 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        /*if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }*/
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -96,6 +120,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        /*if (Gate::denies('delete-post', $post)) {
+            abort(403, "You can't delete this blog post!");
+        }*/
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
